requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
alembic
python-dotenv
pydantic_settings